Separated username validator in a method.
Created a method named `Ranking\Entity\User::getNameValidator()`
that returns the username validator used in the setter of the
same class and property.
This can surely be improved by not leting it create an instance
of that validator every single call, we can probably reuse just
one instance in everywhere.
Right now, I have what I want and that is to re-use that guy
into a route routine that has the username in it, this will then
be the user's home screen: /users/tbon3
Sounds pretty nice from where I am right now. ;)