Created Ranking\Entity\Match test case in PHPUnit.
A match is the object that holds information of when players have
played, whom they played with, where (the map) they played.
Currently I am creating a match with only one *creator*, but a
game to be valid (in - I hope - a near future) should be created
or corroborated/moderated by more than 50% of its players.