Created basic controller for /users/*/matches/teams.
The purpose of this controller is to set up players/users into
valid teams. Supporting many modes of game play with a limit of
10 players (TA's current limit), like: FFA, 1x1, 2x2, 1x2, 2x3.
This controller could be repeatedly called (not implemented) and
sort out if the teams are already defined or if needs more
interaction. Each call to that route should bring the teams close
to a solution.
This work can be easily done by the machine now, but in the future
would be nice to allow all the players already registered to
interfere and validate their team on matches created by another
player.