Created Ranking\Domain\Transaction\UsersToTeams.
This *transaction* should receive users (**addUser**) and race
information about that user and create a team considering the
race chosen, the user and a team number that is always different
for every player (from the 10 possible), supposing a FFA game
match style.

The *Scriptable* interface should be used later on by a different
class thet will run all possible transactions in a pre-defined
order of sulution.

Each *transaction* is a small step towards the game style solution.
Suppose we have this teams in a game:
- tbon3, racx and no-mercy
- dands, sukkoy, zinato
Another step should solve the problem of defining the same team
number for players on the same team so in the end we would have:
- Team 1: tbon3, racx and no-mercy
- Team 2: dands, sukkoy and zinato
Note that the quantity of players in each team should not be
controlled in any form other than user input.